feat(lib): authentication and protected queries for vendors
-added authentication
-protected vendor queries

+++ koa-crud/src/graphql/resolvers.ts
@@ -15,10 +15,18 @@ import {
 } from '../use-cases/members';
 
 const Query = {
-  vendors: async () =>
-    selectAllVendorsUseCase({ id: null, info: null, source: null }),
-  vendor: async (obj, args) =>
-    selectOneVendorUseCase({ id: args.id, info: null, source: null }),
+  vendors: async (obj, args, ctx) => {
+    if (!ctx.allowed) {
+      throw new Error('Forbidden');
+    }
+    return selectAllVendorsUseCase({ id: null, info: null, source: null });
+  },
+  vendor: async (obj, args, ctx) => {
+    if (!ctx.allowed) {
+      throw new Error('Forbidden');
+    }
+    return selectOneVendorUseCase({ id: args.id, info: null, source: null });
+  },
   members: async () =>
     selectAllMembersUseCase({ id: null, info: null, source: null }),
   member: async (obj, args) =>
@@ -26,16 +34,30 @@ const Query = {
 };
 
 const Mutation = {
-  createVendor: async (obj, args) =>
-    insertVendorUseCase({ id: null, info: args.input, source: null }),
-  updateVendor: async (obj, args) =>
-    updateVendorUseCase({
+  createVendor: async (obj, args, ctx) => {
+    if (!ctx.allowed) {
+      throw new Error('Forbidden');
+    }
+
+    return insertVendorUseCase({ id: null, info: args.input, source: null });
+  },
+  updateVendor: async (obj, args, ctx) => {
+    if (!ctx.allowed) {
+      throw new Error('Forbidden');
+    }
+
+    return updateVendorUseCase({
       id: args.input.id,
       info: args.input,
       source: null,
-    }),
-  deleteVendor: async (obj, args) =>
-    deleteOneVendorUseCase({ id: args.id, info: null, source: null }),
+    });
+  },
+  deleteVendor: async (obj, args, ctx) => {
+    if (!ctx.allowed) {
+      throw new Error('Forbidden');
+    }
+    return deleteOneVendorUseCase({ id: args.id, info: null, source: null });
+  },
   createMember: async (obj, args) =>
     insertMemberUseCase({ id: null, info: args.input, source: null }),
   updateMember: async (obj, args) =>
